Add language system | Fixed events
Language :
Using the "i18n" package. Its configuration is located in the file "utils / language.js", and is called inside "index.js" using the function "configure_i18n ();".

The language settings (locale, languageName, abbreviation, dateFormat, aliases) are in the "utils/languages-config.js" file, and the translations are in the "languages/.." folder.

Events :
- "eventName" was getting the wrong parameter.